refactor: Refactored StorageService
