+++ score-client/src/main/java/bio/overture/score/client/command/RepositoryAccessCommand.java
@@ -17,10 +17,10 @@
  */
 package bio.overture.score.client.command;
 
-import java.io.IOException;
-
 import bio.overture.score.client.storage.StorageService;
 import org.springframework.beans.factory.annotation.Autowired;
+
+import java.io.IOException;
 
 public abstract class RepositoryAccessCommand extends AbstractClientCommand {
