+++ score-client/src/main/java/bio/overture/score/client/download/DownloadService.java
@@ -23,7 +23,7 @@ import bio.overture.score.client.exception.NotRetryableException;
 import bio.overture.score.client.exception.RetryableException;
 import bio.overture.score.client.metadata.Entity;
 import bio.overture.score.client.progress.Progress;
-import bio.overture.score.client.storage.StorageServiceInterface;
+import bio.overture.score.client.storage.StorageService;
 import bio.overture.score.client.transport.Transport;
 import bio.overture.score.core.model.ObjectSpecification;
 import bio.overture.score.core.model.Part;
@@ -69,7 +69,7 @@ public class DownloadService {
    * Dependencies.
    */
   @Autowired
-  private StorageServiceInterface storageService;
+  private StorageService storageService;
   @Autowired
   private DownloadStateStore downloadStateStore;
   @Autowired
