+++ score-client/src/main/java/bio/overture/score/client/transport/SequentialPartObjectTransport.java
@@ -19,7 +19,7 @@ package bio.overture.score.client.transport;
 
 import bio.overture.score.client.progress.Progress;
 import bio.overture.score.client.progress.ProgressDataChannel;
-import bio.overture.score.client.storage.StorageServiceInterface;
+import bio.overture.score.client.storage.StorageService;
 import bio.overture.score.core.model.DataChannel;
 import bio.overture.score.core.model.Part;
 import lombok.SneakyThrows;
@@ -36,7 +36,7 @@ import static com.google.common.base.Preconditions.checkNotNull;
 @Slf4j
 public class SequentialPartObjectTransport implements Transport {
 
-  final private StorageServiceInterface proxy;
+  final private StorageService proxy;
   final private Progress progress;
   final private List<Part> parts;
   final private String objectId;
