+++ score-client/src/main/resources/application.yml
@@ -83,6 +83,23 @@ client:
 ---
 
 ###############################################################################
+# Profile - "gen3"
+###############################################################################
+
+spring.profiles: gen3
+
+# Storage server
+storage:
+  url: https://gen3.com
+
+
+client:
+  ssl:
+    custom: false
+
+---
+
+###############################################################################
 # Profile - "collab"
 ###############################################################################
